backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "AI Legal Document Analysis Backend"
authors = ["Jonathan Hop <yithril@gmail.com>"]
readme = "README.md"
packages = [{include = "core"}, {include = "infrastructure"}, {include = "features"}, {include = "services"}, {include = "controllers"}, {include = "dtos"}, {include = "prompts"}, {include = "repositories"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"

# Core Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
python-multipart = "^0.0.12"

# Database (Tortoise ORM + PostgreSQL)
tortoise-orm = "^0.21.0"
aerich = "^0.7.2"
asyncpg = "^0.29.0"

# Storage (MinIO S3-compatible)
minio = "^7.2.0"
aioboto3 = "^13.0.0"

# Vector Database (Pinecone)
pinecone-client = "^5.0.0"

# Search (Elasticsearch)
elasticsearch = "^8.11.0"

# AI/ML (Local Models)
ollama = "^0.3.0"
sentence-transformers = "^3.0.0"

# Document Processing
pymupdf = "^1.24.0"
python-docx = "^1.1.0"
python-magic-bin = "^0.4.14"  # File type detection (Windows-compatible)

# Utilities
python-dotenv = "^1.0.0"
aiofiles = "^24.0.0"
pytesseract = "^0.3.13"
torch = "^2.9.0"
transformers = "^4.57.1"
accelerate = "^1.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "core.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
